requirements.txt
discord.py
python-dotenv
azure-cosmos